requirements.txt
pygad==2.16.0
tqdm==4.62.2
Pillow==8.3.2
pandas==1.3.3
